server/requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
python-multipart==0.0.20
PyMuPDF==1.26.5
Pillow==12.0.0
requests==2.32.5
python-dotenv==1.1.1
pinecone==7.3.0
assemblyai==0.21.0
deepgram-sdk>=3.2.0
google-genai
cloudinary
mistralai
prisma
python-jose
passlib[bcrypt]